refactor: extract CLI business logic into a service layer (item, doc, board) (#50)
* refactor(item): extract CLI business logic into services/items

Move query-building, validation, and column-value rendering out of
src/mondo/cli/item.py into a new src/mondo/services/items.py. The Typer
callbacks keep argument parsing, emission, polling, and exit-code mapping;
the service holds pure helpers (split_filter_expr, build_filter_rule,
parse_columns_csv, build_query_params, parse_column_mapping,
read_batch_input, can_slim_column_values, PositionRelative) and
client-backed helpers (fetch_column_defs, build_column_values,
build_create_variables_for_row, resolve_item_target).

Behavior-neutral: function bodies moved verbatim (renamed _private ->
public), the CLI keeps its try/except -> exit-code mapping, and
GlobalOpts/MondayClient are TYPE_CHECKING-only imports to avoid a cycle.
Drop imports orphaned by the move and fix two doc-comment cross-references
(output/fields.py, subitem.py) that pointed at the relocated symbols.

item.py: 1320 -> 1027 lines. Full unit suite (1639) green; ruff/mypy clean.

* refactor(doc): extract CLI business logic into services/docs

Move block inspection and object-id/doc-id resolution out of
src/mondo/cli/doc.py into a new src/mondo/services/docs.py, mirroring the
item.py extraction. Moved (renamed _private -> public): last_block_id,
resolve_doc_id_from_object_id, object_id_hint_with_client, object_id_hint.
Emit/exit-coupled helpers and all @app.command callbacks stay in the CLI.

Behavior-neutral: bodies moved verbatim, call sites rewired, GlobalOpts/
MondayClient are TYPE_CHECKING-only imports. doc.py: 1272 -> 1213 lines.
Full unit suite (1639) green; ruff/format/mypy clean.

* refactor(board): extract CLI business logic into services/boards

Move type-filter matching, payload decoding, and the live items_count
helpers out of src/mondo/cli/board.py into a new src/mondo/services/boards.py,
mirroring the item.py extraction. Moved (renamed _private -> public):
type_matches, decode_json_string_payload, projection_wants_items_count,
fetch_items_count, plus the BoardTypeFilter enum and its server-value map.
Emit/exit-coupled helpers and all @app.command callbacks stay in the CLI.

Renaming _decode_json_string_payload -> decode_json_string_payload made the
output-contract matrix's source annotation stale; update the generator
(scripts/generate_output_contract_matrix.py) and regenerate the two docs.

Behavior-neutral: bodies moved verbatim, call sites rewired, GlobalOpts is a
TYPE_CHECKING-only import. board.py: 1007 -> 929 lines. Full unit suite
(1639) green; ruff/format/mypy clean.

* fix(subitem): resolve tag names on create + dedup column-value logic (#51)

* fix(subitem): resolve tag names on create + dedup column-value logic

Replace subitem.py's local _build_column_values / _parse_settings copies
with the shared mondo.services.items.build_column_values, removing the
last duplicate of the column-value codec pipeline.

This fixes a latent bug: subitem's copy never resolved 'tags' column
values from names to ids, so 'mondo subitem create --subitems-board <id>
--column <tags-col>=name' failed with 'tags codec accepts integer IDs
only'. The shared service resolves names via create_or_get_tag, matching
'mondo item create'. All other codecs (status/people/...) and raw
passthrough are unchanged.

Add a regression test for tag-name resolution (red before, green after).
subitem.py: 435 -> 374 lines. Full unit suite 1640 green; ruff/format/
mypy clean.

* test(subitem): add live coverage for tag-name resolution on create

Live regression for the subitem tags fix (commit efc2091): exercises
'subitem create --subitems-board <id> --column <tags>=name' end-to-end
against monday.com — the name resolves via create_or_get_tag and lands
on the subitem. The old path exited 5, so a successful create alone
proves the fix; the round-trip confirms the tag landed.
