+++ docs/output-contract-matrix.json
@@ -215,7 +215,7 @@
       "<updated board metadata?>"
     ],
     "notes": "Emits monday's update_board payload directly. Legacy stringified JSON is parsed before emit; non-JSON scalars pass through unchanged.",
-    "source": "_decode_json_string_payload(data.get(\"update_board\"))"
+    "source": "decode_json_string_payload(data.get(\"update_board\"))"
   },
   {
     "command": "cache clear",
